Confirm to play works properly, added createPokemonMoves(pokemonObject) function to create the moves as objects of the pokemon object
js\FR.js
* the confirm popup now works properly. The game is only started if the
player confirms that he is ready. And if the player does not confirm
this, he can click on the start button again to start the game when he
is ready.
* removed redundant comments

js\FR_PKMN.js
* added function createPokemonMoves(pokemonObject). moveTwo, moveThree
and moveFour are now also objects of the pokemon-object of the
player-object, even though they are not properly created if the move
name is ""
* added informative comments about the variables
* removed redundant comments

js\test.js
* added JavaScript-file for testing purposes

pokemonRed.html
* added link to js\test.js for testing purposes